refactor: minify `server` index
