+++ package.json
@@ -8,7 +8,7 @@
     "dev:server": "NELDERMEAD_ENV=dev ts-node ./src/server/index.ts",
     "dev:client": "NELDERMEAD_ENV=dev vite",
     "dev": "NELDERMEAD_ENV=dev concurrently \"npm run dev:server\" \"npm run dev:client\"",
-    "build": "tsc && vite build",
+    "build": "npm run backend:build && tsc && vite build",
     "prod:electron": "electron ./dist/main/index.js",
     "prod:react": "vite preview --port 3000",
     "prod:express": "node ./dist/server/index.js",
